[refactor] add utils package
