+++ test/integration.test.ts
@@ -1,6 +1,6 @@
 import {getPaths} from "./features/paths";
 import {printCollectionToFile} from "./utils/printCollectionToFile";
-import {Shape} from "../src/Shape";
+import {Shape} from "../src/utils/Shape";
 import {clusterPaths} from "../src/zones/clusterPaths";
 
 describe('Integration', () => {
